rewrite of pi_build_modifier gem
