+++ pi_build_modifier/lib/pi_build_modifier/modify/configs/templates/07-resize-init.diff.erb
