--- pi_build_modifier/lib/pi_build_modifier/modify/net-tweaks/templates/wpa_supplicant.conf.erb
@@ -1,5 +1,0 @@
-country=<%= @wpa_country %>
-ctrl_interface=DIR=/var/run/wpa_supplicant GROUP=netdev
-update_config=1
-
-<% @networks.each do |network|  %>network=<%= network %><% end %>
