--- pi_build_modifier/lib/pi_build_modifier/modify/sys_tweaks/templates/ssh.sh.erb
@@ -1,8 +1,0 @@
-#!/bin/bash -e
-
-<% if @enable != 'disable' %>
-on_chroot << EOF
-systemctl <%= @enable %> ssh
-EOF
-
-<% end %>
