+++ pi_build_modifier/lib/pi_build_modifier/version.rb
@@ -19,5 +19,5 @@
 # SOFTWARE.
 
 module PiBuildModifier
-  VERSION = '0.4.0.pre.alpha'
+  VERSION = '0.5.0.pre.alpha'
 end
