--- pi_build_modifier/spec/pi_build_modifier/modify/modifier_cmd_spec.rb
@@ -1,46 +1,0 @@
-# Copyright (c) 2017-2019 Beate Ottenwälder
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-require_relative '../../spec_helper'
-require 'pi_build_modifier/modify/modifier_cmd'
-require 'pi_build_modifier/modify/modifier/modifiers'
-
-module PiBuildModifier
-
-  RSpec.describe Cmd::Modifier do
-
-    context 'initialize' do
-
-      it 'creates an array of modifiers' do
-        modifier = Cmd::Modifier.new("no.cfg", ".")
-        expect(modifier.modifiers).not_to be_nil
-        expect(modifier.modifiers.config_modifiers.length).to be > 0
-      end
-
-      it 'creates erb and config file modifiers' do
-        modifier = Cmd::Modifier.new("no.cfg", ".")
-        expect(modifier.modifiers.config_modifiers).to include(be_a ERBMapper)
-        expect(modifier.modifiers.config_modifiers).to include(be_a ConfigFileModifier)
-      end
-
-    end
-  end
-
-end
